+++ src/de/attributes.rs
@@ -152,7 +152,17 @@ impl<'de> MapAccess<'de> for AttributesDeserializer<'de> {
     {
         match self.value.take() {
             Some(value) => {
-                let de = SimpleTypeDeserializer::from_attr(&value, 0..value.len(), self.version);
+                let value_str = match value {
+                    Cow::Borrowed(b) => Cow::Borrowed(
+                        std::str::from_utf8(b).expect("attribute values are valid UTF-8"), // temporary-#963
+                    ),
+                    Cow::Owned(b) => {
+                        Cow::Owned(String::from_utf8(b).expect("attribute values are valid UTF-8"))
+                        // temporary-#963
+                    }
+                };
+                let de =
+                    SimpleTypeDeserializer::from_attr(&value_str, 0..value_str.len(), self.version);
                 seed.deserialize(de)
             }
             None => Err(DeError::KeyNotRead),