+++ src/reader/buffered_reader.rs
@@ -565,7 +565,8 @@ impl<R: BufRead> Reader<R> {
         // usize (because otherwise we panic at appending to the buffer before that point)
         let end = start + len as usize;
 
-        Ok(BytesText::wrap(&buf[start..end]))
+        let text = std::str::from_utf8(&buf[start..end]).expect("read_text buffer is valid UTF-8"); // temporary-#963
+        Ok(BytesText::wrap(text))
     }
 }